Swap I/O ports some more
The JS emulator has two I/O ports (0 and 1). With this change, the ports
behave a little more like the SIO ports on an 8080-based computer. Port
0 is the status port, which emits 0x20 when input is available, and 0x02
when output is possible. Port 1 is the data port, writing goes to the
console, and reading comes from the keyboard.

With this change, early programs that depend on a teletype input (like
Tiny BASIC) will work properly, and it should be possible to port other
software without tremendous change.