Disable local echo
Previously, when the user typed a character, the JavaScript code would
immediately echo it back to the xterm. Now, this doesn't happen, as it
is replaced by 8080 software. The BIOS (input routine) for Tiny BASIC
now does the echo.

This is useful for other 8080 programs, which perform the echo
themselves (or perhaps optionally turn it off for secret input).