Preserve Palo Alto Tiny BASIC
Palo Alto Tiny BASIC was probably the fourth BASIC published in DDJ.
Unlike TBX (the first), it does not use an interpreter internally. This
version was widely distributed.

It is included here to end up with a more comprehensive collection of
very early 8080 Tiny BASICs. Ideally I can find the source code for TBX
and preserve it here too.